refactor : List<Car> 를 CarGroup 으로 매핑
