+++ src/main/java/racing/domain/Car.java
@@ -11,7 +11,7 @@ package racing.domain;
  */
 public class Car {
 
-  private String name;
+  private final String name;
   private int position;
 
   private final int MAX_NAME_LENGTH = 5;
