+++ src/main/java/racing/domain/RacingGame.java
@@ -1,39 +1,25 @@
 package racing.domain;
 
-import java.util.List;
 import racing.util.RandomNumberGenerator;
 
 public class RacingGame {
 
-  private final List<Car> cars;
+  private final CarGroup carGroup;
 
   public RacingGame(String carsName) {
-    cars = CarFactory.createCars(carsName);
+    carGroup = new CarGroup(carsName);
   }
 
   /**
    * 게임을 한번 실행 한 경우를 의미한다.
    */
-  public List<Car> play() {
-    for (Car car : cars) {
-      car.move(RandomNumberGenerator.generate());
-    }
-    return cars;
+  public CarGroup play() {
+    carGroup.moveCars(RandomNumberGenerator.generate());
+    return carGroup;
   }
 
   public String[] findWinners() {
-    int maxPosition = findMaxPosition();
-    return cars.stream()
-        .filter(car -> car.position() == maxPosition)
-        .map(Car::name)
-        .toArray(String[]::new);
-
+    return carGroup.findWinners();
   }
 
-  private int findMaxPosition() {
-    return cars.stream()
-        .mapToInt(Car::position)
-        .max()
-        .orElse(0);
-  }
 }
